# 01-doc/04-openapi-spec-v2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trading Journal + Prop Firm Simulator API",
    "description": "REST API for the Trading Journal & Prop Firm Challenge Simulator platform. All protected endpoints require Bearer JWT authentication. Financial values are returned as strings to preserve decimal precision.",
    "version": "1.0.0",
    "contact": {
      "name": "Engineering Team",
      "email": "engineering@tradejournal.io"
    }
  },
  "servers": [
    { "url": "https://api.tradejournal.io/v1", "description": "Production" },
    {
      "url": "https://api.staging.tradejournal.io/v1",
      "description": "Staging"
    },
    { "url": "http://localhost:8080/v1", "description": "Local Development" }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /auth/login"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["status", "message"],
        "properties": {
          "status": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "Validation failed" },
          "error": {
            "type": "string",
            "example": "email: must be a valid email address",
            "description": "Optional error detail string. Omitted when empty."
          }
        }
      },
      "PageMetadata": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "size": { "type": "integer", "example": 50 },
          "total_items": { "type": "integer", "example": 248 },
          "total_pages": { "type": "integer", "example": 5 }
        }
      },
...
      "SimulationStatus": {
        "type": "string",
        "enum": ["ACTIVE", "PASSED", "FAILED", "PAUSED"]
      },
      "DrawdownType": {
        "type": "string",
        "enum": ["STATIC", "DAILY", "TRAILING", "COMBINED"]
      },
      "SimulationRules": {
        "type": "object",
        "required": ["drawdown_type", "profit_target_pct", "min_trading_days"],
        "properties": {
          "drawdown_type": { "$ref": "#/components/schemas/DrawdownType" },
          "max_drawdown_pct": {
            "type": "string",
            "example": "10.0000",
            "nullable": true
          },
          "daily_drawdown_pct": {
            "type": "string",
            "example": "5.0000",
            "nullable": true
          },
          "trailing_drawdown_pct": {
            "type": "string",
            "example": "10.0000",
            "nullable": true
          },
          "profit_target_pct": { "type": "string", "example": "10.0000" },
          "min_trading_days": {
            "type": "integer",
            "minimum": 0,
            "example": 10
          },
          "consistency_rule_enabled": { "type": "boolean", "example": false },
          "consistency_threshold_pct": {
            "type": "string",
            "example": "30.0000",
            "nullable": true
          },
          "daily_reset_timezone": { "type": "string", "example": "UTC" },
          "daily_reset_time": { "type": "string", "example": "00:00:00" }
        }
      },
      "SimulationSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "FTMO $100K Phase 1" },
          "account_size": { "type": "string", "example": "100000.00000000" },
          "current_equity": { "type": "string", "example": "102500.00000000" },
          "currency": { "type": "string", "example": "USD" },
          "status": { "$ref": "#/components/schemas/SimulationStatus" },
          "started_at": { "type": "string", "format": "date" },
          "passed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "failed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "total_trades": { "type": "integer" },
          "trading_days_completed": { "type": "integer" },
          "pnl_pct": { "type": "string", "example": "2.5000" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "parameters": {
...
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "PerPageParam": {
        "name": "size",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        }
      }
    },
    "responses": {
      "401Unauthorized": {
        "description": "Missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "status": false,
              "message": "Unauthorized",
              "error": "Invalid or expired token"
            }
          }
        }
      },
...
      "422ValidationError": {
        "description": "Request body validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "status": false,
              "message": "Validation failed",
              "error": "email: must be a valid email address; password: must be at least 8 characters"
            }
          }
        }
      }
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/simulations": {
      "get": {
        "tags": ["Simulations"],
        "summary": "List user simulations",
        "operationId": "listSimulations",
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/PerPageParam" },
          {
            "name": "status",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/SimulationStatus" },
            "description": "Filter by status"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "created_at",
                "-created_at",
                "name",
                "-name",
                "equity",
                "-equity"
              ],
              "default": "-created_at"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated simulation list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "boolean", "example": true },
                    "message": {
                      "type": "string",
                      "example": "Simulations retrieved"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SimulationSummary"
                      }
                    },
                    "metadata": { "$ref": "#/components/schemas/PageMetadata" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401Unauthorized" }
        }
      },
      "post": {
        "tags": ["Simulations"],
        "summary": "Create a new simulation",
        "operationId": "createSimulation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "account_size", "rules"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 150
                  },
                  "account_size": { "type": "string", "example": "100000.00" },
                  "currency": { "type": "string", "default": "USD" },
                  "started_at": { "type": "string", "format": "date" },
                  "notes": { "type": "string", "nullable": true },
                  "template_id": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true,
                    "description": "If provided, rules are copied from template and the 'rules' field can be omitted"
                  },
                  "rules": { "$ref": "#/components/schemas/SimulationRules" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Simulation created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "boolean", "example": true },
                    "message": {
                      "type": "string",
                      "example": "Simulation created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "simulation_id": { "type": "string", "format": "uuid" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401Unauthorized" },
          "422": { "$ref": "#/components/responses/422ValidationError" }
        }
      }
    },
...
  }
}